인증 수정 및 게시물에 따른 사용자 인증 추가
자신이 올린 게시물만 수정하게 인증을 받고, 게시물이 삭제되면 사용자에 참조되어 있는 게시물을 삭제함

+++ Node.js/The Complete Guide/Section_24/controllers/feed.js
@@ -122,6 +122,12 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 422;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        console.log(post.creator.toString(), req.userId);
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) clearImage(post.imageUrl);
       post.title = title;
       post.content = content;
@@ -147,11 +153,22 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 422;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
       res.status(200).json({ message: "Post delete!" });
     })
     .catch((err) => {
